mysql-mcp/pyproject.toml
[project]
name = "mysql-mcp"
version = "0.1.0"
description = "MySQL MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [ "mcp>=1.0.0","aiomysql>=0.2.0","pydantic>=2.10.2","pydantic_settings>=2.6.1","python-dotenv>=1.0.1","uvicorn>=0.32.1","fastapi>=0.115.5","anthropic>=0.40.0",]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mysql-mcp = "mysql_mcp:main"
